rearranged folders, client side and server side
